+++ blueprint/iam-base/domain/mappers/organization.mappers.ts
@@ -1,6 +1,6 @@
 import { schemaValidator } from '@forklaunch/blueprint-core';
 import { requestMapper, responseMapper } from '@forklaunch/core/mappers';
-import { EntityManager, InferEntity, wrap } from '@mikro-orm/core';
+import { EntityManager } from '@mikro-orm/core';
 import { Organization } from '../../persistence/entities/organization.entity';
 import { OrganizationStatus } from '../enum/organizationStatus.enum';
 import { OrganizationSchemas } from '../schemas';
@@ -40,11 +40,10 @@ export const OrganizationMapper = responseMapper({
   schema: OrganizationSchemas.OrganizationSchema(OrganizationStatus),
   entity: Organization,
   mapperDefinition: {
-    toDto: async (entity: InferEntity<typeof Organization>) => {
-      const pojo = wrap(entity).toPOJO();
+    toDto: async (entity) => {
       return {
-        ...pojo,
-        logoUrl: pojo.logoUrl || undefined,
+        ...entity,
+        logoUrl: entity.logoUrl || undefined,
         users: await Promise.all(
           (entity.users.isInitialized()
             ? entity.users