+++ blueprint/iam-base/registrations.ts
@@ -139,8 +139,16 @@ const runtimeDependencies = environmentConfig.chain({
   EntityManager: {
     lifetime: Lifetime.Scoped,
     type: EntityManager,
-    factory: ({ MikroORM }, _resolve, context) =>
-      MikroORM.em.fork(context?.entityManagerOptions as ForkOptions | undefined)
+    factory: (
+      { MikroORM },
+      context: { entityManagerOptions?: ForkOptions; tenantId?: string }
+    ) => {
+      const em = MikroORM.em.fork(context.entityManagerOptions);
+      if (context.tenantId) {
+        em.setFilterParams('tenant', { tenantId: context.tenantId });
+      }
+      return em;
+    }
   }
 });
 
@@ -154,10 +162,10 @@ const serviceDependencies = runtimeDependencies.chain({
       OrganizationMapperTypes,
       OrganizationDtoTypes
     >,
-    factory: ({ EntityManager, OpenTelemetryCollector }, resolve, context) =>
+    factory: ({ EntityManager, OpenTelemetryCollector }, context, resolve) =>
       new BaseOrganizationService(
-        context.entityManagerOptions
-          ? resolve('EntityManager', context)
+        context?.entityManagerOptions
+          ? resolve?.('EntityManager', context)
           : EntityManager,
         OpenTelemetryCollector,
         schemaValidator,
@@ -175,7 +183,7 @@ const serviceDependencies = runtimeDependencies.chain({
       PermissionMapperTypes,
       PermissionDtoTypes
     >,
-    factory: ({ EntityManager, OpenTelemetryCollector }, resolve, context) =>
+    factory: ({ EntityManager, OpenTelemetryCollector }, context, resolve) =>
       new BasePermissionService(
         context.entityManagerOptions
           ? resolve('EntityManager', context)
@@ -194,7 +202,7 @@ const serviceDependencies = runtimeDependencies.chain({
   RoleService: {
     lifetime: Lifetime.Scoped,
     type: BaseRoleService<SchemaValidator, RoleMapperTypes, RoleDtoTypes>,
-    factory: ({ EntityManager, OpenTelemetryCollector }, resolve, context) =>
+    factory: ({ EntityManager, OpenTelemetryCollector }, context, resolve) =>
       new BaseRoleService(
         context.entityManagerOptions
           ? resolve('EntityManager', context)
@@ -216,7 +224,7 @@ const serviceDependencies = runtimeDependencies.chain({
       UserMapperTypes,
       UserDtoTypes
     >,
-    factory: ({ EntityManager, OpenTelemetryCollector }, resolve, context) =>
+    factory: ({ EntityManager, OpenTelemetryCollector }, context, resolve) =>
       new BaseUserService(
         EntityManager,
         () => resolve('RoleService', context),