+++ framework/validator/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @forklaunch/validator
 
+## 1.2.1
+
+### Patch Changes
+
+- fix compliance entity
+- Updated dependencies
+  - @forklaunch/common@1.2.1
+
+## 1.2.0
+
+### Minor Changes
+
+- Validator 25% performance uptick and cleaner Config Injector syntax
+
+### Patch Changes
+
+- Updated dependencies
+  - @forklaunch/common@1.2.0
+
 ## 1.1.8
 
 ### Patch Changes